refactor upgrade menu UI to use floating text service for upgrade notifications
