+++ Salvager/Assets/Scripts/Installer/LevelSelectorInstaller.cs
@@ -28,7 +28,7 @@ namespace Installer
             
             // UI
             Container.Bind<IPanelManagerUI>().To<PanelManagerUI>().FromComponentsInHierarchy().AsSingle();
-            Container.Bind<IOnCanvasFloatingTextUI>().To<OnCanvasFloatingTextUI>().FromComponentsInHierarchy().AsSingle();
+            Container.Bind<IFloatingTextService>().To<FloatingTextService>().FromComponentsInHierarchy().AsSingle();
         }        
     }
 }
